refactor: split jira.rs and daemon.rs under the 500-line file cap
Both grew past CLAUDE.md's 500-line limit. Extract Jira's three
status-transition writers (close/reopen/cancel, which share one GET
transitions -> pick -> POST shape) into jira_transitions.rs, and the
tray's pause/resume commands (pause_for_duration/pause_indefinitely/
resume_capture + their clock/label helpers) into commands/pause.rs.
No behavior change — pure extraction, tests moved with their subjects.
