+++ tray/src-tauri/src/commands.rs
@@ -37,6 +37,7 @@ pub mod notices;
 pub mod notifications;
 pub mod openobserve;
 pub mod parents;
+pub mod pause;
 pub mod settings;
 pub mod setup;
 pub mod system;
@@ -58,6 +59,7 @@ pub use notices::*;
 pub use notifications::*;
 pub use openobserve::*;
 pub use parents::*;
+pub use pause::*;
 pub use settings::*;
 pub use setup::*;
 pub use system::*;
